Implement role and user management with permission checks, enhance navigation links, and update route definitions for better organization and access control.

+++ app/Livewire/Admin/Pages/Article/Category.php
@@ -4,13 +4,14 @@ namespace App\Livewire\Admin\Pages\Article;
 
 use App\Livewire\Forms\ArticleCategoryForm;
 use App\Models\ArticleCategory;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Attributes\Layout;
 use Livewire\Component;
 use Livewire\WithPagination;
 
 class Category extends Component
 {
-    use WithPagination;
+    use WithPagination, AuthorizesRequests;
 
     #[Layout('layouts.app')]
 
@@ -21,6 +22,7 @@ class Category extends Component
 
     public function create()
     {
+        $this->authorize('access', 'admin.articles.categories.create');
         $this->form->reset();
         $this->editId = null;
         $this->deleteId = null;
@@ -29,6 +31,7 @@ class Category extends Component
 
     public function edit($id)
     {
+        $this->authorize('access', 'admin.articles.categories.edit');
         $this->editId = $id;
         $category = ArticleCategory::find($id);
         $this->form->setCategory($category);
@@ -38,10 +41,12 @@ class Category extends Component
     public function save()
     {
         if ($this->editId) {
+            $this->authorize('access', 'admin.articles.categories.edit');
             $this->form->update();
             $this->editId = null;
             toastr()->success('Category updated successfully');
         } else {
+            $this->authorize('access', 'admin.articles.categories.create');
             $this->form->store();
             toastr()->success('Category created successfully');
         }
@@ -50,6 +55,7 @@ class Category extends Component
 
     public function confirmDelete($id)
     {
+        $this->authorize('access', 'admin.articles.categories.delete');
         $this->deleteId = $id;
         $category = ArticleCategory::find($id);
         $this->form->setCategory($category);
@@ -58,6 +64,7 @@ class Category extends Component
 
     public function delete()
     {
+        $this->authorize('access', 'admin.articles.categories.delete');
         if ($this->deleteId) {
             $this->form->delete();
             $this->deleteId = null;
@@ -68,6 +75,8 @@ class Category extends Component
 
     public function render()
     {
+        $this->authorize('access', 'admin.articles.categories.index');
+
         $table_heads = ['#', 'Name', 'Actions'];
 
         $categories = ArticleCategory::when($this->search, function ($query) {
