+++ app/Livewire/Admin/Pages/Article/Form.php
@@ -6,6 +6,7 @@ use App\Livewire\Forms\ArticleForm;
 use App\Models\Article;
 use App\Models\ArticleCategory;
 use App\Models\Tag;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Support\Str;
 use Livewire\Attributes\Layout;
 use Livewire\Component;
@@ -14,7 +15,7 @@ use Livewire\WithPagination;
 
 class Form extends Component
 {
-    use WithPagination, WithFileUploads;
+    use WithPagination, WithFileUploads, AuthorizesRequests;
 
     #[Layout('layouts.app')]
 
@@ -27,8 +28,11 @@ class Form extends Component
 
     public function mount($id = null)
     {
+        $this->authorize('access', 'admin.articles.create');
+
         if ($id) {
-            // Jika edit
+            $this->authorize('access', 'admin.articles.edit');
+
             $article = Article::find($id);
             $this->form->setArticle($article);
             $this->editId = $article->id;
@@ -60,9 +64,14 @@ class Form extends Component
     public function save()
     {
         if ($this->editId) {
+            $this->authorize('access', 'admin.articles.edit');
+
             $this->form->is_published = true;
             $this->form->update();
+            toastr()->success('Artikel berhasil diperbarui dan dipublikasikan');
         } else {
+            $this->authorize('access', 'admin.articles.create');
+
             $this->form->is_published = true;
             $this->form->store();
             $this->form->reset();
@@ -75,17 +84,21 @@ class Form extends Component
     public function preview()
     {
         $this->previewContent = $this->form->content;
-        // dd($this->previewContent);
+
         $this->dispatch('open-modal', 'preview-modal');
     }
 
     public function confirmDelete()
     {
+        $this->authorize('access', 'admin.articles.destroy');
+
         $this->dispatch('open-modal', 'delete-article-confirmation');
     }
 
     public function delete()
     {
+        $this->authorize('access', 'admin.articles.destroy');
+
         $this->form->delete();
         $this->form->reset();
         toastr()->success('Artikel berhasil dihapus');
@@ -94,6 +107,8 @@ class Form extends Component
 
     public function forceDelete()
     {
+        $this->authorize('access', 'admin.articles.destroy');
+
         $this->form->forceDelete();
         $this->form->reset();
         toastr()->success('Artikel berhasil dihapus permanen');
@@ -105,18 +120,25 @@ class Form extends Component
         $this->form->is_published = false;
 
         if ($this->editId) {
+            $this->authorize('access', 'admin.articles.edit');
+
             $this->form->update();
+            toastr()->success('Artikel berhasil diperbarui sebagai draft');
         } else {
+            $this->authorize('access', 'admin.articles.create');
+
             $this->form->store();
+            toastr()->success('Artikel berhasil disimpan sebagai draft');
         }
 
         $this->form->reset();
-        toastr()->success('Artikel berhasil disimpan sebagai draft');
         return redirect()->route('admin.articles.index');
     }
 
     public function unpublish()
     {
+        $this->authorize('access', 'admin.articles.unpublish');
+
         $this->form->unpublish();
         $this->form->reset();
         toastr()->success('Artikel berhasil di unpublish');
