+++ app/Livewire/Admin/Pages/Article/Tag.php
@@ -4,13 +4,14 @@ namespace App\Livewire\Admin\Pages\Article;
 
 use App\Livewire\Forms\TagForm;
 use App\Models\Tag as ModelsTag;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Attributes\Layout;
 use Livewire\Component;
 use Livewire\WithPagination;
 
 class Tag extends Component
 {
-    use WithPagination;
+    use WithPagination, AuthorizesRequests;
 
     #[Layout('layouts.app')]
 
@@ -21,6 +22,8 @@ class Tag extends Component
 
     public function create()
     {
+        $this->authorize('access', 'admin.articles.tags.create');
+
         $this->form->reset();
         $this->editId = null;
         $this->deleteId = null;
@@ -29,6 +32,8 @@ class Tag extends Component
 
     public function edit($id)
     {
+        $this->authorize('access', 'admin.articles.tags.edit');
+
         $this->editId = $id;
         $tag = ModelsTag::find($id);
         $this->form->setTag($tag);
@@ -38,18 +43,25 @@ class Tag extends Component
     public function save()
     {
         if ($this->editId) {
+            $this->authorize('access', 'admin.articles.tags.edit');
+
             $this->form->update();
             $this->editId = null;
             toastr()->success('Tag updated successfully');
         } else {
+            $this->authorize('access', 'admin.articles.tags.create');
+
             $this->form->store();
             toastr()->success('Tag created successfully');
         }
+
         $this->dispatch('close-modal', 'tag-modal');
     }
 
     public function confirmDelete($id)
     {
+        $this->authorize('access', 'admin.articles.tags.destroy');
+
         $this->deleteId = $id;
         $tag = ModelsTag::find($id);
         $this->form->setTag($tag);
@@ -59,6 +71,8 @@ class Tag extends Component
 
     public function delete()
     {
+        $this->authorize('access', 'admin.articles.tags.destroy');
+
         if ($this->deleteId) {
             $this->form->delete();
             $this->deleteId = null;
@@ -68,8 +82,13 @@ class Tag extends Component
     }
     public function render()
     {
+        // Authorization
+        $this->authorize('access', 'admin.articles.tags.index');
+
+        // Table heads
         $table_heads = ['#', 'Name', 'Actions'];
 
+        // Table data
         $tags = ModelsTag::when($this->search, function ($query) {
             $query->where('name', 'like', '%' . $this->search . '%');
         })->latest()->paginate(5);
