+++ app/Livewire/Admin/Pages/Asset/Category.php
@@ -4,13 +4,14 @@ namespace App\Livewire\Admin\Pages\Asset;
 
 use App\Livewire\Forms\AssetCategoryForm;
 use App\Models\AssetCategory;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Attributes\Layout;
 use Livewire\Component;
 use Livewire\WithPagination;
 
 class Category extends Component
 {
-    use WithPagination;
+    use WithPagination, AuthorizesRequests;
 
     #[Layout('layouts.app')]
 
@@ -19,8 +20,15 @@ class Category extends Component
     public $editId = null;
     public $deleteId = null;
 
+    public function updatedFormName()
+    {
+        $this->form->slug = str($this->form->name)->slug();
+    }
+
     public function create()
     {
+        $this->authorize('access', 'admin.asset-categories.create');
+
         $this->form->reset();
         $this->editId = null;
         $this->deleteId = null;
@@ -29,6 +37,8 @@ class Category extends Component
 
     public function edit($id)
     {
+        $this->authorize('access', 'admin.asset-categories.edit');
+
         $this->editId = $id;
         $category = AssetCategory::find($id);
         $this->form->setCategory($category);
@@ -38,10 +48,14 @@ class Category extends Component
     public function save()
     {
         if ($this->editId) {
+            $this->authorize('access', 'admin.asset-categories.edit');
+
             $this->form->update();
             $this->editId = null;
             toastr()->success('Category updated successfully');
         } else {
+            $this->authorize('access', 'admin.asset-categories.create');
+
             $this->form->store();
             toastr()->success('Category created successfully');
         }
@@ -50,6 +64,8 @@ class Category extends Component
 
     public function confirmDelete($id)
     {
+        $this->authorize('access', 'admin.asset-categories.destroy');
+
         $this->deleteId = $id;
         $category = AssetCategory::find($id);
         $this->form->setCategory($category);
@@ -58,20 +74,28 @@ class Category extends Component
 
     public function delete()
     {
+        $this->authorize('access', 'admin.asset-categories.destroy');
+
         if ($this->deleteId) {
             $this->form->delete();
             $this->deleteId = null;
             toastr()->success('Category deleted successfully');
         }
+
         $this->dispatch('close-modal', 'delete-category-confirmation');
     }
     public function render()
     {
-        $table_heads = ['#', 'Name', 'Status', 'Actions'];
+        // Authorization
+        $this->authorize('access', 'admin.asset-categories.index');
+
+        $table_heads = ['#', 'Name', 'slug', 'Status', 'Actions'];
 
         $categories = AssetCategory::when($this->search, function ($query) {
-            $query->where('name', 'like', '%' . $this->search . '%');
+            $query->where('name', 'like', '%' . $this->search . '%')
+                ->orWhere('slug', 'like', '%' . $this->search . '%');
         })->latest()->paginate(5);
+
         return view('livewire.admin.pages.asset.category', [
             'categories' => $categories,
             'table_heads' => $table_heads,
