+++ app/Livewire/Admin/Pages/Borrowing/Index.php
@@ -5,13 +5,14 @@ namespace App\Livewire\Admin\Pages\Borrowing;
 use App\Enums\BorrowingStatus;
 use App\Livewire\Forms\BorrowingForm;
 use App\Models\Borrowing;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Attributes\Layout;
 use Livewire\Component;
 use Livewire\WithPagination;
 
 class Index extends Component
 {
-    use WithPagination;
+    use WithPagination, AuthorizesRequests;
 
     #[Layout('layouts.app')]
 
@@ -26,6 +27,8 @@ class Index extends Component
 
     public function show(Borrowing $borrowing)
     {
+        $this->authorize('access', 'admin.asset-borrowings.show');
+
         $this->reset('borrowing');
         $this->borrowing = $borrowing->load(['assets', 'user', 'event']);
         // dd($this->borrowing->toArray());
@@ -35,18 +38,24 @@ class Index extends Component
 
     public function confirmApprove($id)
     {
+        $this->authorize('access', 'admin.asset-borrowings.approve');
+
         $this->approveId = $id;
         $this->dispatch('open-modal', 'approve-borrowing-confirmation');
     }
 
     public function confirmReject($id)
     {
+        $this->authorize('access', 'admin.asset-borrowings.reject');
+
         $this->rejectId = $id;
         $this->dispatch('open-modal', 'reject-borrowing-confirmation');
     }
 
     public function confirmDelete(Borrowing $borrowing)
     {
+        $this->authorize('access', 'admin.asset-borrowings.delete');
+
         // dd($borrowing);
         $this->deleteId = $borrowing->id;
         $this->borrowing = $borrowing;
@@ -55,6 +64,8 @@ class Index extends Component
 
     public function approve()
     {
+        $this->authorize('access', 'admin.asset-borrowings.approve');
+
         if ($this->approveId) {
             $this->form->approve($this->approveId);
 
@@ -71,6 +82,8 @@ class Index extends Component
 
     public function reject()
     {
+        $this->authorize('access', 'admin.asset-borrowings.reject');
+
         if ($this->rejectId) {
             Borrowing::find($this->rejectId)->update(['status' => BorrowingStatus::REJECTED]);
             $this->rejectId = null;
@@ -82,6 +95,8 @@ class Index extends Component
 
     public function destroy()
     {
+        $this->authorize('access', 'admin.asset-borrowings.destroy');
+
         if ($this->deleteId) {
             $this->form->setBorrowing(Borrowing::find($this->deleteId));
             $this->form->destroy();
@@ -94,6 +109,8 @@ class Index extends Component
 
     public function render()
     {
+        $this->authorize('access', 'admin.asset-borrowings.index');
+
         $table_heads = ['#', 'Borrower', 'Event', 'date', 'Note', 'Status', 'Actions'];
 
         $borrowings = Borrowing::with(['assets', 'user', 'event'])->latest()->paginate(5);
