+++ app/Livewire/Admin/Pages/Event/Show.php
@@ -7,11 +7,14 @@ use App\Models\Event;
 use App\Models\EventCategory;
 use App\Models\Location;
 use App\Models\Organization;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Attributes\Layout;
 use Livewire\Component;
 
 class Show extends Component
 {
+    use AuthorizesRequests;
+
     #[Layout('layouts.app')]
 
     public EventShowForm $form;
@@ -22,6 +25,8 @@ class Show extends Component
 
     public function mount(Event $event)
     {
+        $this->authorize('access', 'admin.events.show');
+
         $this->event = $event;
 
         $this->categories = EventCategory::active()->get(['id', 'name']);
@@ -31,18 +36,24 @@ class Show extends Component
 
     public function confirmEdit($id)
     {
+        $this->authorize('access', 'admin.events.edit');
+
         $this->form->setEvent($this->event);
         $this->dispatch('open-modal', 'update-event-confirmation');
     }
 
     public function edit()
     {
+        $this->authorize('access', 'admin.events.edit');
+
         $this->dispatch('close-modal', 'update-event-confirmation');
         $this->dispatch('open-modal', 'event-modal');
     }
 
     public function save()
     {
+        $this->authorize('access', 'admin.events.edit');
+
         $this->form->update();
 
         // Update data tanpa redirect
@@ -54,12 +65,16 @@ class Show extends Component
 
     public function confirmDelete()
     {
+        $this->authorize('access', 'admin.events.destroy');
+
         $this->form->setEvent($this->event);
         $this->dispatch('open-modal', 'delete-event-confirmation');
     }
 
     public function delete()
     {
+        $this->authorize('access', 'admin.events.destroy');
+
         $this->form->delete();
         $this->dispatch('deleteSuccess');
         $this->redirect(route('admin.events.index'));
@@ -67,6 +82,8 @@ class Show extends Component
 
     public function render()
     {
+        $this->authorize('access', 'admin.events.show');
+
         return view('livewire.admin.pages.event.show');
     }
 }
