+++ app/Livewire/Admin/Pages/EventCategory.php
@@ -4,13 +4,14 @@ namespace App\Livewire\Admin\Pages;
 
 use App\Livewire\Forms\EventCategoryForm;
 use App\Models\EventCategory as ModelsEventCategory;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Attributes\Layout;
 use Livewire\Component;
 use Livewire\WithPagination;
 
 class EventCategory extends Component
 {
-    use WithPagination;
+    use WithPagination, AuthorizesRequests;
 
     #[Layout('layouts.app')]
 
@@ -21,6 +22,8 @@ class EventCategory extends Component
 
     public function create()
     {
+        $this->authorize('access', 'admin.event-categories.create');
+
         $this->form->reset();
         $this->editId = null;
         $this->deleteId = null;
@@ -29,6 +32,8 @@ class EventCategory extends Component
 
     public function edit($id)
     {
+        $this->authorize('access', 'admin.event-categories.edit');
+
         $this->editId = $id;
         $category = ModelsEventCategory::find($id);
         $this->form->setCategory($category);
@@ -38,10 +43,14 @@ class EventCategory extends Component
     public function save()
     {
         if ($this->editId) {
+            $this->authorize('access', 'admin.event-categories.edit');
+
             $this->form->update();
             $this->editId = null;
             $this->dispatch('updateSuccess');
         } else {
+            $this->authorize('access', 'admin.event-categories.create');
+
             $this->form->store();
             $this->dispatch('createSuccess');
         }
@@ -50,6 +59,8 @@ class EventCategory extends Component
 
     public function confirmDelete($id)
     {
+        $this->authorize('access', 'admin.event-categories.destroy');
+
         $this->deleteId = $id;
         $category = ModelsEventCategory::find($id);
         $this->form->setCategory($category);
@@ -58,6 +69,8 @@ class EventCategory extends Component
 
     public function delete()
     {
+        $this->authorize('access', 'admin.event-categories.destroy');
+
         if ($this->deleteId) {
             $this->form->delete();
             $this->deleteId = null;
@@ -68,6 +81,8 @@ class EventCategory extends Component
 
     public function render()
     {
+        $this->authorize('access', 'admin.event-categories.index');
+
         $table_heads = ['#', 'Name', 'Color', 'Status', 'Actions'];
 
         $categories = ModelsEventCategory::when($this->search, function ($query) {
