+++ app/Livewire/Admin/Pages/Groups.php
@@ -4,13 +4,14 @@ namespace App\Livewire\Admin\Pages;
 
 use App\Livewire\Forms\GroupForm;
 use App\Models\Group;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Attributes\Layout;
 use Livewire\Component;
 use Livewire\WithPagination;
 
 class Groups extends Component
 {
-    use WithPagination;
+    use WithPagination, AuthorizesRequests;
 
     #[Layout('layouts.app')]
 
@@ -21,6 +22,8 @@ class Groups extends Component
 
     public function create()
     {
+        $this->authorize('access', 'admin.groups.create');
+
         $this->form->reset();
         $this->editId = null;
         $this->deleteId = null;
@@ -29,6 +32,8 @@ class Groups extends Component
 
     public function edit($id)
     {
+        $this->authorize('access', 'admin.groups.edit');
+
         $this->editId = $id;
         $group = Group::find($id);
         $this->form->setGroup($group);
@@ -38,18 +43,24 @@ class Groups extends Component
     public function save()
     {
         if ($this->editId) {
+            $this->authorize('access', 'admin.groups.edit');
+
             $this->form->update();
             $this->editId = null;
-            $this->dispatch('updateSuccess');
+            toastr()->success('Group updated successfully');
         } else {
+            $this->authorize('access', 'admin.groups.create');
+
             $this->form->store();
-            $this->dispatch('createSuccess');
+            toastr()->success('Group created successfully');
         }
         $this->dispatch('close-modal', 'group-modal');
     }
 
     public function confirmDelete($id)
     {
+        $this->authorize('access', 'admin.groups.destroy');
+
         $this->deleteId = $id;
         $group = Group::find($id);
         $this->form->setGroup($group);
@@ -58,10 +69,12 @@ class Groups extends Component
 
     public function delete()
     {
+        $this->authorize('access', 'admin.groups.destroy');
+
         if ($this->deleteId) {
             $this->form->delete();
             $this->deleteId = null;
-            $this->dispatch('deleteSuccess');
+            toastr()->success('Group deleted successfully');
         }
         $this->dispatch('close-modal', 'delete-group-confirmation');
     }
@@ -69,6 +82,8 @@ class Groups extends Component
 
     public function render()
     {
+        $this->authorize('access', 'admin.groups.index');
+
         $table_heads = ['#', 'Name', 'Actions'];
 
         $groups = Group::when($this->search, function ($query) {
