+++ app/Livewire/Admin/Pages/Location.php
@@ -4,6 +4,8 @@ namespace App\Livewire\Admin\Pages;
 
 use App\Livewire\Forms\LocationForm;
 use App\Models\Location as ModelsLocation;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use Illuminate\Support\Facades\Log;
 use Livewire\Attributes\Layout;
 use Livewire\Component;
 use Livewire\WithFileUploads;
@@ -11,7 +13,7 @@ use Livewire\WithPagination;
 
 class Location extends Component
 {
-    use WithPagination, WithFileUploads;
+    use WithPagination, WithFileUploads, AuthorizesRequests;
 
     #[Layout('layouts.app')]
 
@@ -20,8 +22,30 @@ class Location extends Component
     public $editId = null;
     public $deleteId = null;
 
+
+    public function updatedFormImage()
+    {
+        Log::info('Upload masuk', [
+            'token_match' => request()->session()->token() === request()->header('X-CSRF-TOKEN'),
+            'session_exists' => session()->has('_token'),
+            'csrf_token' => request()->session()->token(),
+            'file_sementara' => [
+                'originalName' => $this->form->image->getClientOriginalName(),
+                'size' => $this->form->image->getSize(),
+                'mime' => $this->form->image->getMimeType(),
+            ]
+        ]);
+    }
+
+    public function updatedActivated()
+    {
+        $this->form->is_active = $this->activated;
+    }
+
     public function create()
     {
+        $this->authorize('access', 'admin.locations.create');
+
         $this->form->reset();
         $this->editId = null;
         $this->deleteId = null;
@@ -30,6 +54,8 @@ class Location extends Component
 
     public function edit($id)
     {
+        $this->authorize('access', 'admin.locations.edit');
+
         $this->editId = $id;
         $location = ModelsLocation::find($id);
         $this->form->setLocation($location);
@@ -39,18 +65,23 @@ class Location extends Component
     public function save()
     {
         if ($this->editId) {
+            $this->authorize('access', 'admin.locations.edit');
             $this->form->update();
             $this->editId = null;
-            $this->dispatch('updateSuccess');
+            toastr()->success('Location updated successfully');
         } else {
+            $this->authorize('access', 'admin.locations.create');
+
             $this->form->store();
-            $this->dispatch('createSuccess');
+            toastr()->success('Location created successfully');
         }
         $this->dispatch('close-modal', 'location-modal');
     }
 
     public function confirmDelete($id)
     {
+        $this->authorize('access', 'admin.locations.destroy');
+
         $this->deleteId = $id;
         $location = ModelsLocation::find($id);
         $this->form->setLocation($location);
@@ -59,6 +90,8 @@ class Location extends Component
 
     public function delete()
     {
+        $this->authorize('access', 'admin.locations.destroy');
+
         if ($this->deleteId) {
             $this->form->delete();
             $this->deleteId = null;
@@ -74,6 +107,8 @@ class Location extends Component
 
     public function render()
     {
+        $this->authorize('access', 'admin.locations.index');
+
         $table_heads = ['#', 'Image', 'Name', 'Description', 'Status', 'Actions'];
 
         $locations = ModelsLocation::when($this->search, function ($query) {
