+++ app/Livewire/Admin/Pages/Organization.php
@@ -4,13 +4,14 @@ namespace App\Livewire\Admin\Pages;
 
 use App\Livewire\Forms\OrganizationForm;
 use App\Models\Organization as ModelsOrganization;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Attributes\Layout;
 use Livewire\Component;
 use Livewire\WithPagination;
 
 class Organization extends Component
 {
-    use WithPagination;
+    use WithPagination, AuthorizesRequests;
 
     #[Layout('layouts.app')]
 
@@ -21,6 +22,8 @@ class Organization extends Component
 
     public function create()
     {
+        $this->authorize('access', 'admin.organizations.create');
+
         $this->form->reset();
         $this->editId = null;
         $this->deleteId = null;
@@ -29,6 +32,8 @@ class Organization extends Component
 
     public function edit($id)
     {
+        $this->authorize('access', 'admin.organizations.edit');
+
         $this->editId = $id;
         $organization = ModelsOrganization::find($id);
         $this->form->setOrganization($organization);
@@ -38,18 +43,24 @@ class Organization extends Component
     public function save()
     {
         if ($this->editId) {
+            $this->authorize('access', 'admin.organizations.edit');
+
             $this->form->update();
             $this->editId = null;
-            $this->dispatch('updateSuccess');
+            toastr()->success('Organization updated successfully');
         } else {
+            $this->authorize('access', 'admin.organizations.create');
+
             $this->form->store();
-            $this->dispatch('createSuccess');
+            toastr()->success('Organization created successfully');
         }
         $this->dispatch('close-modal', 'organization-modal');
     }
 
     public function confirmDelete($id)
     {
+        $this->authorize('access', 'admin.organizations.destroy');
+
         $this->deleteId = $id;
         $organization = ModelsOrganization::find($id);
         $this->form->setOrganization($organization);
@@ -58,16 +69,20 @@ class Organization extends Component
 
     public function delete()
     {
+        $this->authorize('access', 'admin.organizations.destroy');
+
         if ($this->deleteId) {
             $this->form->delete();
             $this->deleteId = null;
-            $this->dispatch('deleteSuccess');
+            toastr()->success('Organization deleted successfully');
         }
         $this->dispatch('close-modal', 'delete-organization-confirmation');
     }
 
     public function render()
     {
+        $this->authorize('access', 'admin.organizations.index');
+
         $table_heads = ['#', 'Name', 'Code', 'Description', 'Status', 'Actions'];
 
         $organizations = ModelsOrganization::when($this->search, function ($query) {
